매개변수 id 타입 int -> long 수정

+++ base_java/miniproject_todolist/repository/TodoRepository.java
@@ -8,9 +8,9 @@ public interface TodoRepository {
 
     void editTodo(TodoItem item);
 
-    void deleteTodo(int id);
+    void deleteTodo(long id);
 
     List<TodoItem> getAllTodos();
 
-    TodoItem getTodoById(int id);
+    TodoItem getTodoById(long id);
 }
